Added Koding guide
This was written after installation, if someone wouldn't mind testing it
to make sure I covered everything, that would be great. It's similar to
ubuntu, except I got a package error when trying to install npm and
nodejs at the same time. So splitting them into chris-leas package and
then nodejs seemed to resolve this.