Update apache proxies configuration for 2.4.X
The current solution isn't working for the latest(?) apache version 2.4.7 on Ubuntu. Would be nice to get some others test it on different OS and other apache 2.4.x and maybe even 2.5.X versions before this is merged.

For reference reported on the forums by [whitts](https://community.nodebb.org/topic/7890/websocket-error-in-console/10).