Proxy configuration tutorial for node.js
This is a tutorial to use node.js as a reverse proxy, with websockets,
TLS, and proxy rules. It may be easier for some people than using nginx,
if they are already familiar with node.js

I have a working example that uses this method running at
support.globlee.com/forum, on a AWS free tier instance, using mongo,
and it seems to be doing fine.